test: split duplicate proof tests
