+++ test/clawsweeper.test.ts
@@ -1239,2043 +1239,6 @@ if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$
 `;
 }
 
-test("apply-decisions ignores bare refs inside cross-repo markdown link labels for duplicate proof", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 357,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by [PR #400](https://github.com/other/repo/pull/400)",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by the linked external PR.",
-      ),
-      357,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "357.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 357,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Unrelated same-repo PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Unrelated provider cleanup.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "failure",
-            message: "coverage proof should not run for cross-repo markdown link labels",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    assert.equal(existsSync(proofLogPath), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions ignores bare refs inside same-repo markdown link labels for duplicate proof", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 358,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by [fixes #123](https://github.com/openclaw/openclaw/pull/400)",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by the linked canonical PR.",
-      ),
-      358,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "358.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 358,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          123: {
-            number: 123,
-            title: "Unrelated draft PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/123",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            draft: true,
-            body: "Unrelated provider cleanup.",
-            comments: [],
-            labels: [],
-          },
-          400: {
-            number: 400,
-            title: "Canonical provider replacement",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Covers the provider route fallback behavior.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-            invocationLogPath: proofLogPath,
-            expectedPromptIncludes: "Canonical provider replacement",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-    }>;
-    assert.equal(
-      report.find((entry) => entry.number === 358)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(readFileSync(proofLogPath, "utf8"), /proof/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions keeps newline-start bare PR refs tied to their own line", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 358,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify(["Links:\n#400 supersedes this PR and #500 is related"]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by the linked canonical PR.",
-      ),
-      358,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "358.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 358,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-        linkedIssues: {
-          500: {
-            number: 500,
-            title: "Related provider issue",
-            html_url: "https://github.com/openclaw/openclaw/issues/500",
-            state: "open",
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(
-      report.find((entry) => entry.number === 358)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(readFileSync(proofLogPath, "utf8"), /proof/);
-    assert.match(
-      report.find((entry) => entry.number === 358)?.reason ?? "",
-      /unique fallback route behavior/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions ignores unrelated same-line bare PR refs for duplicate proof", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 361,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify(["Superseded by #400; #500 is related"]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by the linked canonical PR.",
-      ),
-      361,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "361.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 361,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 361.",
-            comments: [],
-            labels: [],
-          },
-          500: {
-            number: 500,
-            title: "Related draft PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/500",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            draft: true,
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-      JSON.stringify(report),
-    );
-    assert.match(
-      report.find((entry) => entry.action === "closed")?.reason ?? "",
-      /duplicate or superseded/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-for (const scenario of [
-  { name: "and", number: 364, refs: "#400 and #401" },
-  { name: "comma", number: 365, refs: "#400, #401" },
-  { name: "semicolon", number: 366, refs: "#400; #401" },
-  { name: "comma and and", number: 367, refs: "#400, #401, and #402" },
-  {
-    name: "markdown links",
-    number: 368,
-    refs: "[first](https://github.com/openclaw/openclaw/pull/400), [second](https://github.com/openclaw/openclaw/pull/401)",
-  },
-]) {
-  test(`apply-decisions preserves supersession context across shorthand PR ref lists with ${scenario.name}`, () => {
-    const root = mkdtempSync(tmpPrefix);
-    try {
-      const itemsDir = join(root, "items");
-      const closedDir = join(root, "closed");
-      const plansDir = join(root, "plans");
-      const reportPath = join(root, "apply-report.json");
-      mkdirSync(itemsDir, { recursive: true });
-      mkdirSync(plansDir, { recursive: true });
-      const synced = reportWithSyncedReviewComment(
-        lowSignalCloseReport({
-          number: scenario.number,
-          title: "Provider route fallback",
-          close_reason: "duplicate_or_superseded",
-          work_cluster_refs: JSON.stringify([`Superseded by ${scenario.refs}`]),
-        }).replace(
-          "Closing this PR because the branch is not a useful landing base.",
-          "Closing this PR as superseded by the linked canonical PRs.",
-        ),
-        scenario.number,
-        "duplicate_or_superseded",
-      );
-      writeFileSync(join(itemsDir, `${scenario.number}.md`), synced.report, "utf8");
-
-      withMockGh(
-        root,
-        promotionGhMock({
-          number: scenario.number,
-          title: "Provider route fallback",
-          comment: synced.comment,
-          linkedPulls: {
-            400: {
-              number: 400,
-              title: "Initial provider cleanup",
-              html_url: "https://github.com/openclaw/openclaw/pull/400",
-              state: "closed",
-              merged_at: "2026-05-01T00:00:00Z",
-              body: "Cleans up provider setup without changing the fallback route.",
-              comments: [],
-              labels: [],
-            },
-            401: {
-              number: 401,
-              title: "Provider cleanup",
-              html_url: "https://github.com/openclaw/openclaw/pull/401",
-              state: "closed",
-              merged_at: "2026-05-02T00:00:00Z",
-              body: `Includes the fallback route behavior from PR ${scenario.number}.`,
-              comments: [],
-              labels: [],
-            },
-            402: {
-              number: 402,
-              title: "Later provider cleanup",
-              html_url: "https://github.com/openclaw/openclaw/pull/402",
-              state: "closed",
-              merged_at: "2026-05-03T00:00:00Z",
-              body: "Follow-up provider cleanup.",
-              comments: [],
-              labels: [],
-            },
-          },
-        }),
-        () => {
-          withMockCodexProof(
-            root,
-            {
-              type: "decision",
-              decision: "keep_open",
-              reason: "PR B carries forward PR A's fallback route behavior.",
-              keepOpenPromptIncludes:
-                "Cleans up provider setup without changing the fallback route.",
-              coveredPromptIncludes: `Includes the fallback route behavior from PR ${scenario.number}.`,
-            },
-            () => {
-              runApplyDecisionsForTest({
-                itemsDir,
-                closedDir,
-                plansDir,
-                reportPath,
-                extraArgs: [
-                  "--target-repo",
-                  "openclaw/openclaw",
-                  "--dry-run",
-                  "--apply-kind",
-                  "all",
-                  "--processed-limit",
-                  "3",
-                ],
-              });
-            },
-          );
-        },
-      );
-
-      const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-        action: string;
-        reason: string;
-      }>;
-      assert.equal(
-        report.some((entry) => entry.action === "closed"),
-        true,
-        JSON.stringify(report),
-      );
-      assert.match(
-        report.find((entry) => entry.action === "closed")?.reason ?? "",
-        /duplicate or superseded/,
-      );
-    } finally {
-      rmSync(root, { recursive: true, force: true });
-    }
-  });
-}
-
-test("apply-decisions does not proof-gate duplicate PR closes with bare issue refs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 357,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify(["Duplicate of #456"]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as a duplicate of canonical issue #456.",
-      ),
-      357,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "357.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 357,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedIssues: {
-          456: {
-            number: 456,
-            title: "Provider fallback tracker",
-            html_url: "https://github.com/openclaw/openclaw/issues/456",
-            state: "open",
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.equal(JSON.stringify(report).includes("proof should not run"), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions preserves full PR URL evidence over later bare refs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 348,
-        title: "Already proposed duplicate close",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.\nLater issue discussion also mentions #400.",
-      ),
-      348,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "348.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 348,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedIssues: {
-          400: {
-            number: 400,
-            title: "Issue with same number",
-            html_url: "https://github.com/openclaw/openclaw/issues/400",
-            state: "open",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "failure",
-            message: "proof should not run",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "kept_open")?.reason ?? "",
-      /linked canonical PR #400 could not be read/,
-    );
-    assert.equal(existsSync(proofLogPath), false);
-    assert.equal(existsSync(join(closedDir, "348.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions checks duplicate PR coverage proof before syncing corrected review comments", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const commentWriteLogPath = join(root, "comment-writes.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const reportMarkdown = lowSignalCloseReport({
-      number: 354,
-      title: "Unsynced duplicate close",
-      close_reason: "duplicate_or_superseded",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    }).replace(
-      "Closing this PR because the branch is not a useful landing base.",
-      "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-    );
-    writeFileSync(join(itemsDir, "354.md"), reportMarkdown, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 354,
-        title: "Unsynced duplicate close",
-        comment: "",
-        comments: [],
-        issueCommentCount: 0,
-        commentWriteLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "review_comment_synced"),
-      false,
-    );
-    assert.equal(
-      report.find((entry) => entry.number === 354)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(
-      report.find((entry) => entry.number === 354)?.reason ?? "",
-      /unique fallback route behavior/,
-    );
-    assert.match(readFileSync(commentWriteLogPath, "utf8"), /issues\/354\/comments/);
-    const blockedReport = readFileSync(join(itemsDir, "354.md"), "utf8");
-    assert.match(blockedReport, /^decision: keep_open$/m);
-    assert.match(blockedReport, /^review_comment_synced_at:/m);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions closes existing duplicate PR close proposals when coverage proof says covered", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 349,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      349,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "349.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 349,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 349.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "closed")?.reason ?? "",
-      /duplicate or superseded/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions records successful duplicate PR coverage proof for closed PRs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const closeAppliedBodyLogPath = join(root, "close-applied-body.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 360,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      360,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "360.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 360,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        closeAppliedBodyLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 360.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const closedReport = readFileSync(join(closedDir, "360.md"), "utf8");
-    assert.match(closedReport, /## PR Close Coverage Proof/);
-    assert.match(closedReport, /Reason: PR B carries forward PR A's fallback route behavior\./);
-    const closeAppliedBody = readFileSync(closeAppliedBodyLogPath, "utf8");
-    assert.match(
-      closeAppliedBody,
-      /Coverage proof: PR B carries forward PR A's fallback route behavior\./,
-    );
-    assert.match(
-      closeAppliedBody,
-      /Covering PR: https:\/\/github\.com\/openclaw\/openclaw\/pull\/400\./,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions filters covering PR bot comments from coverage proof", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 363,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      363,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "363.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 363,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 363.",
-            comments: [
-              {
-                id: 9400,
-                html_url: "https://github.com/openclaw/openclaw/pull/400#issuecomment-9400",
-                created_at: "2026-05-01T02:00:00Z",
-                updated_at: "2026-05-01T02:00:00Z",
-                user: { login: "clawsweeper[bot]" },
-                body: [
-                  "AUTOMATION_SHOULD_NOT_REACH_SWEEP_PROOF",
-                  "",
-                  "<!-- clawsweeper-review item=400 -->",
-                ].join("\n"),
-              },
-              {
-                id: 9401,
-                html_url: "https://github.com/openclaw/openclaw/pull/400#issuecomment-9401",
-                created_at: "2026-05-01T03:00:00Z",
-                updated_at: "2026-05-01T03:00:00Z",
-                user: { login: "maintainer" },
-                body: "HUMAN_COVERING_CONTEXT_REACHES_SWEEP_PROOF",
-              },
-            ],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            expectedPromptIncludes: "HUMAN_COVERING_CONTEXT_REACHES_SWEEP_PROOF",
-            unexpectedPromptIncludes: "AUTOMATION_SHOULD_NOT_REACH_SWEEP_PROOF",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions rechecks duplicate PR freshness after coverage proof passes", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 357,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      357,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "357.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 357,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        itemUpdatedAtAfterProof: "2026-05-01T00:05:00Z",
-        itemUpdatedAtAfterProofLogPath: proofLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 357.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(report[0]?.action, "skipped_changed_since_review");
-    assert.match(report[0]?.reason ?? "", /updated_at changed/);
-    assert.equal(existsSync(join(closedDir, "357.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions rechecks covering PR freshness after coverage proof passes", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 360,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      360,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "360.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 360,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        itemUpdatedAtAfterProofLogPath: proofLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            updated_at: "2026-05-01T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 360.",
-            comments: [],
-            labels: [],
-          },
-        },
-        linkedPullsAfterProof: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            updated_at: "2026-05-01T00:05:00Z",
-            body: "Changed after proof ran.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "retry_pr_close_coverage_proof")?.reason ?? "",
-      /linked canonical PR #400 changed after coverage proof/,
-    );
-    assert.equal(existsSync(join(closedDir, "360.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions allows self-synced labels after proof with truncated context", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    const labelLogPath = join(root, "labels.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 359,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        labels: JSON.stringify(["status: 📣 needs proof"]),
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      359,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "359.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 359,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        issueCommentCount: 25,
-        itemUpdatedAtAfterLabelSync: "2026-05-01T00:04:00Z",
-        itemUpdatedAtAfterLabelSyncLogPath: labelLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 359.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.match(readFileSync(labelLogPath, "utf8"), /issue edit 359/);
-    assert.ok(existsSync(join(closedDir, "359.md")));
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions blocks post-proof human activity hidden by self-updates", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    const labelLogPath = join(root, "labels.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 362,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        labels: JSON.stringify(["status: 📣 needs proof"]),
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      362,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "362.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 362,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        comments: [
-          {
-            id: 9362,
-            html_url: "https://github.com/openclaw/openclaw/pull/362#issuecomment-9362",
-            created_at: "2099-01-01T00:00:00Z",
-            updated_at: "2099-01-01T00:00:00Z",
-            user: { login: "contributor" },
-            body: "Please do not close this yet.",
-          },
-        ],
-        itemUpdatedAtAfterLabelSync: "2026-05-01T00:04:00Z",
-        itemUpdatedAtAfterLabelSyncLogPath: labelLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 362.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(report[0]?.action, "skipped_changed_since_review");
-    assert.match(report[0]?.reason ?? "", /non-automation activity after coverage proof/);
-    assert.match(readFileSync(labelLogPath, "utf8"), /issue edit 362/);
-    assert.equal(existsSync(join(closedDir, "362.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions keeps existing duplicate PR close proposals open when coverage proof fails", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 350,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      350,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "350.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 350,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "May or may not include PR 350.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "model unavailable" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "retry_pr_close_coverage_proof")?.reason ?? "",
-      /PR close coverage proof failed/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions retries transient duplicate PR coverage proof failures", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 353,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      353,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "353.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 353,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 353.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "failure",
-            message: "temporary model outage",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    let report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "retry_pr_close_coverage_proof")?.reason ?? "",
-      /temporary/,
-    );
-    assert.match(
-      readFileSync(join(itemsDir, "353.md"), "utf8"),
-      /^action_taken: retry_pr_close_coverage_proof$/m,
-    );
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 353,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 353.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B carries forward PR A's fallback route behavior.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.equal(readFileSync(proofLogPath, "utf8").trim().split("\n").length, 2);
-    assert.ok(existsSync(join(closedDir, "353.md")));
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions checks age before duplicate PR coverage proof", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 351,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      351,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "351.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 351,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the fallback route behavior from PR 351.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-              "--min-age-days",
-              "99999",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "kept_open")?.reason ?? "",
-      /created less than or equal to 99999 days ago/,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions ignores unrelated unsafe PR links when canonical PR is safe", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const reportMarkdown = lowSignalCloseReport({
-      number: 347,
-      title: "Already proposed duplicate close",
-      close_reason: "duplicate_or_superseded",
-      work_cluster_refs: JSON.stringify(["https://github.com/openclaw/openclaw/pull/401"]),
-    }).replace(
-      "Closing this PR because the branch is not a useful landing base.",
-      [
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-        "",
-        "Earlier context also mentioned https://github.com/openclaw/openclaw/pull/401.",
-      ].join("\n"),
-    );
-    const synced = reportWithSyncedReviewComment(reportMarkdown, 347, "duplicate_or_superseded");
-    writeFileSync(join(itemsDir, "347.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 347,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Merged canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            labels: [],
-          },
-          401: {
-            number: 401,
-            title: "Unrelated closed PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/401",
-            state: "closed",
-            merged_at: null,
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B is the merged canonical PR covering PR A.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions blocks duplicate close when canonical PR is a bare cluster ref", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 341,
-        title: "Already proposed duplicate close",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify(["https://github.com/openclaw/openclaw/pull/400"]),
-      }),
-      341,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "341.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 341,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Closed unmerged canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: null,
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "kept_open")?.reason ?? "",
-      /closed and unmerged/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions retries duplicate close when linked canonical PR comments cannot be read", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 340,
-        title: "Already proposed duplicate close",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      340,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "340.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 340,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Includes the provider cleanup from PR 340.",
-            comments: [{ body: "temporary hydration target" }],
-            commentsError: "temporary comments outage",
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "retry_pr_close_coverage_proof")?.reason ?? "",
-      /temporary comments outage/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
 test("apply-decisions does not promote PRs superseded by PRs already proposed for close", () => {
   const root = mkdtempSync(tmpPrefix);
   try {
